Add regexp prefix matching
This adds support for matching on regular expressions in addition to
literal strings.

Regular expressions are slightly tedious because -- if they use capture
groups -- they make the offset of the suffix capture group unreliable.
As a result, it's necessary to select only the last capture group as
the forwarded message.

The prefix regexp is stored as an instance variable to avoid generating
it more often than is necessary.